requirements.txt
# EightFold Talent Intelligence - Dependencies

# Core
requests>=2.28.0
PyGithub>=2.1.0
python-dateutil>=2.8.0

# Analysis
numpy>=1.24.0
pandas>=2.0.0

# UI
streamlit>=1.28.0

# LLM Integration (optional - for dynamic JD parsing)
openai>=1.0.0
anthropic>=0.18.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0

